Split image utils (#279)
* Easier maintenance and testing
* Adding some tests for utils files
* Moving tests to own folder
