--- pocs/utils/images.py
@@ -1,1052 +1,0 @@
-import os
-import shutil
-import subprocess
-
-from dateutil import parser as date_parser
-from json import loads
-
-import matplotlib
-matplotlib.use('Agg')
-from matplotlib import pyplot as plt
-from warnings import warn
-
-import numpy as np
-
-from skimage.feature import canny
-from skimage.transform import hough_circle
-from skimage.transform import hough_circle_peaks
-
-from astropy.io import fits
-from astropy.nddata import Cutout2D
-from astropy.visualization import SqrtStretch
-from astropy.visualization.mpl_normalize import ImageNormalize
-from astropy.wcs import WCS
-from astropy import units as u
-
-from ffmpy import FFmpeg
-from glob import glob
-
-from pocs.utils import current_time
-from pocs.utils import error
-from pocs.utils.config import load_config
-
-
-def solve_field(fname, timeout=15, solve_opts=[], **kwargs):
-    """ Plate solves an image.
-
-    Args:
-        fname(str, required):       Filename to solve in either .cr2 or .fits
-                                    extension.
-        timeout(int, optional):     Timeout for the solve-field command,
-                                    defaults to 60 seconds.
-        solve_opts(list, optional): List of options for solve-field.
-        verbose(bool, optional):    Show output, defaults to False.
-    """
-    verbose = kwargs.get('verbose', False)
-    if verbose:
-        print("Entering solve_field")
-
-    solve_field_script = "{}/scripts/solve_field.sh".format(
-        os.getenv('POCS'), '/var/panoptes/POCS')
-
-    if not os.path.exists(solve_field_script):  # pragma: no cover
-        raise error.InvalidSystemCommand(
-            "Can't find solve-field: {}".format(solve_field_script))
-
-    # Add the options for solving the field
-    if solve_opts:
-        options = solve_opts
-    else:
-        options = [
-            '--guess-scale',
-            '--cpulimit', str(timeout),
-            '--no-verify',
-            '--no-plots',
-            '--crpix-center',
-            '--match', 'none',
-            '--corr', 'none',
-            '--wcs', 'none',
-            '--downsample', '4',
-        ]
-
-        if kwargs.get('clobber', True):
-            options.append('--overwrite')
-        if kwargs.get('skip_solved', True):
-            options.append('--skip-solved')
-
-        if 'ra' in kwargs:
-            options.append('--ra')
-            options.append(str(kwargs.get('ra')))
-        if 'dec' in kwargs:
-            options.append('--dec')
-            options.append(str(kwargs.get('dec')))
-        if 'radius' in kwargs:
-            options.append('--radius')
-            options.append(str(kwargs.get('radius')))
-
-    cmd = [solve_field_script] + options + [fname]
-    if verbose:
-        print("Cmd:", cmd)
-
-    try:
-        proc = subprocess.Popen(cmd, universal_newlines=True,
-                                stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-    except OSError as e:
-        raise error.InvalidCommand(
-            "Can't send command to solve_field.sh: {} \t {}".format(e, cmd))
-    except ValueError as e:
-        raise error.InvalidCommand(
-            "Bad parameters to solve_field: {} \t {}".format(e, cmd))
-    except Exception as e:
-        raise error.PanError("Timeout on plate solving: {}".format(e))
-
-    if verbose:
-        print("Returning proc from solve_field")
-
-    return proc
-
-
-def get_solve_field(fname, replace=True, remove_extras=True, **kwargs):
-    """Convenience function to wait for `solve_field` to finish.
-
-    This function merely passes the `fname` of the image to be solved along to `solve_field`,
-    which returns a subprocess.Popen object. This function then waits for that command
-    to complete, populates a dictonary with the EXIF informaiton and returns. This is often
-    more useful than the raw `solve_field` function
-
-    Args:
-        fname ({str}): Name of file to be solved, either a FITS or CR2
-        replace (bool, optional): Replace fname the solved file
-        remove_extras (bool, optional): Remove the files generated by solver
-        **kwargs ({dict}): Options to pass to `solve_field`
-
-    Returns:
-        dict: Keyword information from the solved field
-    """
-    verbose = kwargs.get('verbose', False)
-    out_dict = {}
-    output = None
-    errs = None
-
-    # Check for solved file
-    if kwargs.get('skip_solved', True) and \
-            (os.path.exists(fname.replace('.fits', '.solved')) or WCS(fname).is_celestial):
-        if verbose:
-            print("Solved file exists, skipping",
-                  "(pass skip_solved=False to solve again):",
-                  fname)
-
-        out_dict['solved_fits_file'] = fname
-        return out_dict
-
-    if verbose:
-        print("Entering get_solve_field:", fname)
-
-    # Set a default radius of 15
-    kwargs.setdefault('radius', 15)
-
-    proc = solve_field(fname, **kwargs)
-    try:
-        output, errs = proc.communicate(timeout=kwargs.get('timeout', 30))
-    except subprocess.TimeoutExpired:
-        proc.kill()
-        raise error.Timeout("Timeout while solving")
-    else:
-        if verbose:
-            print("Returncode:", proc.returncode)
-            print("Output:", output)
-            print("Errors:", errs)
-
-        if proc.returncode == 3:
-            raise error.SolveError('solve-field not found: {}'.format(output))
-
-        if not os.path.exists(fname.replace('.fits', '.solved')):
-            raise error.SolveError('File not solved')
-
-        try:
-            # Handle extra files created by astrometry.net
-            new = fname.replace('.fits', '.new')
-            rdls = fname.replace('.fits', '.rdls')
-            axy = fname.replace('.fits', '.axy')
-            xyls = fname.replace('.fits', '-indx.xyls')
-
-            if replace and os.path.exists(new):
-                # Remove converted fits
-                os.remove(fname)
-                # Rename solved fits to proper extension
-                os.rename(new, fname)
-
-                out_dict['solved_fits_file'] = fname
-            else:
-                out_dict['solved_fits_file'] = new
-
-            if remove_extras:
-                for f in [rdls, xyls, axy]:
-                    if os.path.exists(f):
-                        os.remove(f)
-
-        except Exception as e:
-            warn('Cannot remove extra files: {}'.format(e))
-
-    if errs is not None:
-        warn("Error in solving: {}".format(errs))
-    else:
-        # Read the EXIF information from the CR2
-        if fname.endswith('cr2'):
-            out_dict.update(read_exif(fname))
-
-        try:
-            out_dict.update(fits.getheader(fname))
-        except OSError:
-            if verbose:
-                print("Can't read fits header for:", fname)
-
-    return out_dict
-
-
-def crop_data(data, box_width=200, center=None, verbose=False):
-    """ Return a cropped portion of the image
-
-    Shape is a box centered around the middle of the data
-
-    Args:
-        data(np.array):     The original data, e.g. an image.
-        box_width(int):     Size of box width in pixels, defaults to 200px
-        center(tuple(int)): Crop around set of coords, defaults to image center.
-
-    Returns:
-        np.array:           A clipped (thumbnailed) version of the data
-    """
-    assert data.shape[0] >= box_width, "Can't clip data, it's smaller than {} ({})".format(
-        box_width, data.shape)
-    # Get the center
-    if verbose:
-        print("Data to crop: {}".format(data.shape))
-
-    if center is None:
-        x_len, y_len = data.shape
-        x_center = int(x_len / 2)
-        y_center = int(y_len / 2)
-    else:
-        y_center = int(center[0])
-        x_center = int(center[1])
-
-    box_width = int(box_width / 2)
-
-    if verbose:
-        print("Using center: {} {}".format(x_center, y_center))
-        print("Box width: {}".format(box_width))
-
-    center = data[x_center - box_width: x_center + box_width,
-                  y_center - box_width: y_center + box_width]
-
-    return center
-
-
-def get_wcsinfo(fits_fname, verbose=False):
-    """Returns the WCS information for a FITS file.
-
-    Uses the `wcsinfo` astrometry.net utility script to get the WCS information
-    from a plate-solved file.
-
-    Parameters
-    ----------
-    fits_fname : {str}
-        Name of a FITS file that contains a WCS.
-    verbose : {bool}, optional
-        Verbose (the default is False)
-    Returns
-    -------
-    dict
-        Output as returned from `wcsinfo`
-    """
-    assert os.path.exists(fits_fname), warn(
-        "No file exists at: {}".format(fits_fname))
-
-    wcsinfo = shutil.which('wcsinfo')
-    if wcsinfo is None:
-        raise error.InvalidCommand('wcsinfo not found')
-
-    run_cmd = [wcsinfo, fits_fname]
-
-    if verbose:
-        print("wcsinfo command: {}".format(run_cmd))
-
-    proc = subprocess.Popen(run_cmd, stdout=subprocess.PIPE,
-                            stderr=subprocess.STDOUT, universal_newlines=True)
-    try:
-        output, errs = proc.communicate(timeout=5)
-    except subprocess.TimeoutExpired:  # pragma: no cover
-        proc.kill()
-        output, errs = proc.communicate()
-
-    unit_lookup = {
-        'crpix0': u.pixel,
-        'crpix1': u.pixel,
-        'crval0': u.degree,
-        'crval1': u.degree,
-        'cd11': (u.deg / u.pixel),
-        'cd12': (u.deg / u.pixel),
-        'cd21': (u.deg / u.pixel),
-        'cd22': (u.deg / u.pixel),
-        'imagew': u.pixel,
-        'imageh': u.pixel,
-        'pixscale': (u.arcsec / u.pixel),
-        'orientation': u.degree,
-        'ra_center': u.degree,
-        'dec_center': u.degree,
-        'orientation_center': u.degree,
-        'ra_center_h': u.hourangle,
-        'ra_center_m': u.minute,
-        'ra_center_s': u.second,
-        'dec_center_d': u.degree,
-        'dec_center_m': u.minute,
-        'dec_center_s': u.second,
-        'fieldarea': (u.degree * u.degree),
-        'fieldw': u.degree,
-        'fieldh': u.degree,
-        'decmin': u.degree,
-        'decmax': u.degree,
-        'ramin': u.degree,
-        'ramax': u.degree,
-        'ra_min_merc': u.degree,
-        'ra_max_merc': u.degree,
-        'dec_min_merc': u.degree,
-        'dec_max_merc': u.degree,
-        'merc_diff': u.degree,
-    }
-
-    wcs_info = {}
-    for line in output.split('\n'):
-        try:
-            k, v = line.split(' ')
-            try:
-                v = float(v)
-            except Exception:
-                pass
-
-            wcs_info[k] = float(v) * unit_lookup.get(k, 1)
-        except ValueError:
-            pass
-            # print("Error on line: {}".format(line))
-
-    wcs_info['wcs_file'] = fits_fname
-
-    return wcs_info
-
-
-def fpack(fits_fname, unpack=False, verbose=False):
-    """ Compress/Decompress a FITS file
-
-    Uses `fpack` (or `funpack` if `unpack=True`) to compress a FITS file
-
-    Parameters
-    ----------
-    fits_fname : {str}
-        Name of a FITS file that contains a WCS.
-    unpack : {bool}, optional
-        file should decompressed instead of compressed (default is False)
-    verbose : {bool}, optional
-        Verbose (the default is False)
-    Returns
-    -------
-    str
-        Filename of compressed/decompressed file
-    """
-    assert os.path.exists(fits_fname), warn(
-        "No file exists at: {}".format(fits_fname))
-
-    if unpack:
-        fpack = shutil.which('funpack')
-        run_cmd = [fpack, '-D', fits_fname]
-        out_file = fits_fname.replace('.fz', '')
-    else:
-        fpack = shutil.which('fpack')
-        run_cmd = [fpack, '-D', '-Y', fits_fname]
-        out_file = fits_fname.replace('.fits', '.fits.fz')
-
-    try:
-        assert fpack is not None
-    except AssertionError:
-        warn("fpack not found (try installing cfitsio). File has not been changed")
-        return fits_fname
-
-    if verbose:
-        print("fpack command: {}".format(run_cmd))
-
-    proc = subprocess.Popen(run_cmd, stdout=subprocess.PIPE,
-                            stderr=subprocess.STDOUT, universal_newlines=True)
-    try:
-        output, errs = proc.communicate(timeout=5)
-    except subprocess.TimeoutExpired:
-        proc.kill()
-        output, errs = proc.communicate()
-
-    return out_file
-
-
-def make_pretty_image(fname, timeout=15, **kwargs):  # pragma: no cover
-    """ Make a pretty image
-
-    This calls out to an external script which will try to extract the JPG
-    directly from the CR2 file, otherwise will do an actual conversion
-
-    Notes:
-        See `$POCS/scripts/cr2_to_jpg.sh`
-
-    Arguments:
-        fname {str} -- Name of CR2 file
-        **kwargs {dict} -- Additional arguments to be passed to external script
-
-    Keyword Arguments:
-        timeout {number} -- Process timeout (default: {15})
-
-    Returns:
-        str -- Filename of image that was created
-
-    """
-    assert os.path.exists(fname),\
-        warn("File doesn't exist, can't make pretty: {}".format(fname))
-
-    if fname.endswith('.cr2'):
-        return _make_pretty_from_cr2(fname, timeout=timeout, **kwargs)
-    elif fname.endswith('.fits'):
-        return _make_pretty_from_fits(fname, timeout=timeout, **kwargs)
-
-
-def _make_pretty_from_fits(fname, timeout=15, **kwargs):
-    config = load_config()
-
-    title = '{} {}'.format(kwargs.get('title', ''), current_time().isot)
-
-    new_filename = fname.replace('.fits', '.jpg')
-
-    data = fits.getdata(fname)
-    plt.imshow(data, cmap='cubehelix_r', origin='lower')
-    plt.title(title)
-    plt.savefig(new_filename)
-
-    image_dir = config['directories']['images']
-
-    ln_fn = '{}/latest.jpg'.format(image_dir)
-
-    try:
-        os.remove(ln_fn)
-    except FileNotFoundError:
-        pass
-
-    try:
-        os.symlink(new_filename, ln_fn)
-    except Exception as e:
-        warn("Can't link latest image: {}".format(e))
-
-    return new_filename
-
-
-def _make_pretty_from_cr2(fname, timeout=15, **kwargs):  # pragma: no cover
-    verbose = kwargs.get('verbose', False)
-
-    title = '{} {}'.format(kwargs.get('title', ''), current_time().isot)
-
-    solve_field = "{}/scripts/cr2_to_jpg.sh".format(os.getenv('POCS'))
-    cmd = [solve_field, fname, title]
-
-    if kwargs.get('primary', False):
-        cmd.append('link')
-
-    if verbose:
-        print(cmd)
-
-    try:
-        proc = subprocess.Popen(cmd, stdout=subprocess.DEVNULL,
-                                stderr=subprocess.DEVNULL)
-        if verbose:
-            print(proc)
-    except OSError as e:
-        raise error.InvalidCommand("Can't send command to gphoto2."
-                                   " {} \t {}".format(e, cmd))
-    except ValueError as e:
-        raise error.InvalidCommand("Bad parameters to gphoto2."
-                                   " {} \t {}".format(e, cmd))
-    except Exception as e:
-        raise error.PanError("Timeout on plate solving: {}".format(e))
-
-    return fname.replace('cr2', 'jpg')
-
-
-def focus_metric(data, merit_function='vollath_F4', **kwargs):
-    """Compute the focus metric.
-
-    Computes a focus metric on the given data using a supplied merit function.
-    The merit function can be passed either as the name of the function (must be
-    defined in this module) or as a callable object. Additional keyword arguments
-    for the merit function can be passed as keyword arguments to this function.
-
-    Args:
-        data (numpy array) -- 2D array to calculate the focus metric for.
-        merit_function (str/callable) -- Name of merit function (if in
-            pocs.utils.images) or a callable object.
-
-    Returns:
-        scalar: result of calling merit function on data
-    """
-    if isinstance(merit_function, str):
-        try:
-            merit_function = globals()[merit_function]
-        except KeyError:
-            raise KeyError(
-                "Focus merit function '{}' not found in pocs.utils.images!".format(merit_function))
-
-    return merit_function(data, **kwargs)
-
-
-def vollath_F4(data, axis=None):
-    """Compute F4 focus metric
-
-    Computes the F_4 focus metric as defined by Vollath (1998) for the given 2D
-    numpy array. The metric can be computed in the y axis, x axis, or the mean of
-    the two (default).
-
-    Arguments:
-        data (numpy array) -- 2D array to calculate F4 on.
-        axis (str, optional, default None) -- Which axis to calculate F4 in. Can
-            be 'Y'/'y', 'X'/'x' or None, which will calculate the F4 value for
-            both axes and return the mean.
-
-    Returns:
-        float64: Calculated F4 value for y, x axis or both
-    """
-    if axis == 'Y' or axis == 'y':
-        return _vollath_F4_y(data)
-    elif axis == 'X' or axis == 'x':
-        return _vollath_F4_x(data)
-    elif not axis:
-        return (_vollath_F4_y(data) + _vollath_F4_x(data)) / 2
-    else:
-        raise ValueError(
-            "axis must be one of 'Y', 'y', 'X', 'x' or None, got {}!".format(axis))
-
-
-def mask_saturated(data, saturation_level=None, threshold=0.9, dtype=np.float64):
-    if not saturation_level:
-        try:
-            # If data is an integer type use iinfo to compute machine limits
-            dtype_info = np.iinfo(data.dtype)
-        except ValueError:
-            # Not an integer type. Assume for now we have 16 bit data
-            saturation_level = threshold * (2**16 - 1)
-        else:
-            # Data is an integer type, set saturation level at specified fraction of
-            # max value for the type
-            saturation_level = threshold * dtype_info.max
-
-    # Convert data to masked array of requested dtype, mask values above saturation level
-    return np.ma.array(data, mask=(data > saturation_level), dtype=dtype)
-
-
-def _vollath_F4_y(data):
-    A1 = (data[1:] * data[:-1]).mean()
-    A2 = (data[2:] * data[:-2]).mean()
-    return A1 - A2
-
-
-def _vollath_F4_x(data):
-    A1 = (data[:, 1:] * data[:, :-1]).mean()
-    A2 = (data[:, 2:] * data[:, :-2]).mean()
-    return A1 - A2
-
-#######################################################################
-# IO Functions
-#######################################################################
-
-
-def cr2_to_fits(
-        cr2_fname,
-        fits_fname=None,
-        clobber=False,
-        headers={},
-        fits_headers={},
-        remove_cr2=False,
-        **kwargs):  # pragma: no cover
-    """ Convert a CR2 file to FITS
-
-    This is a convenience function that first converts the CR2 to PGM via `cr2_to_pgm`.
-    Also adds keyword headers to the FITS file.
-
-    Note:
-        The intermediate PGM file is automatically removed
-
-    Arguments:
-        cr2_fname {str} -- Name of CR2 file to be converted
-        **kwargs {dict} -- Additional keywords to be used
-
-    Keyword Arguments:
-        fits_fname {str} -- Name of FITS file to output. If None (default), the `cr2_fname` is used
-            as base (default: {None})
-        clobber {bool} -- A bool indicating if existing FITS should be clobbered (default: {False})
-        headers {dict} -- Header data that is filtered and added to the FITS header.
-        fits_headers {dict} -- Header data that is added to the FITS header without filtering.
-        remove_cr2 {bool} -- A bool indicating if the CR2 should be removed (default: {False})
-
-    """
-
-    verbose = kwargs.get('verbose', False)
-    assert os.path.exists(cr2_fname),\
-        warn("File doesn't exist, can't convert cr2 to fits: {}".format(cr2_fname))
-
-    if fits_fname is None:
-        fits_fname = cr2_fname.replace('.cr2', '.fits')
-
-    if not os.path.exists(fits_fname) or clobber:
-        if verbose:
-            print("Converting CR2 to PGM: {}".format(cr2_fname))
-
-        # Convert the CR2 to a PGM file then delete PGM
-        pgm = read_pgm(cr2_to_pgm(cr2_fname), remove_after=True)
-
-        # Add the EXIF information from the CR2 file
-        exif = read_exif(cr2_fname)
-
-        # Set the PGM as the primary data for the FITS file
-        hdu = fits.PrimaryHDU(pgm)
-
-        obs_date = date_parser.parse(
-            exif.get('DateTimeOriginal', '').replace(':', '-', 2)).isoformat()
-
-        # Set some default headers
-        hdu.header.set('FILTER', 'RGGB')
-        hdu.header.set('ISO', exif.get('ISO', ''))
-        hdu.header.set('EXPTIME', exif.get('ExposureTime', 'Seconds'))
-        hdu.header.set('CAMTEMP', exif.get(
-            'CameraTemperature', ''), 'Celsius - From CR2')
-        hdu.header.set('CIRCCONF', exif.get(
-            'CircleOfConfusion', ''), 'From CR2')
-        hdu.header.set('COLORTMP', exif.get(
-            'ColorTempMeasured', ''), 'From CR2')
-        hdu.header.set('FILENAME', exif.get('FileName', ''), 'From CR2')
-        hdu.header.set('INTSN', exif.get(
-            'InternalSerialNumber', ''), 'From CR2')
-        hdu.header.set('CAMSN', exif.get('SerialNumber', ''), 'From CR2')
-        hdu.header.set('MEASEV', exif.get('MeasuredEV', ''), 'From CR2')
-        hdu.header.set('MEASEV2', exif.get('MeasuredEV2', ''), 'From CR2')
-        hdu.header.set('MEASRGGB', exif.get('MeasuredRGGB', ''), 'From CR2')
-        hdu.header.set('WHTLVLN', exif.get('NormalWhiteLevel', ''), 'From CR2')
-        hdu.header.set('WHTLVLS', exif.get(
-            'SpecularWhiteLevel', ''), 'From CR2')
-        hdu.header.set('REDBAL', exif.get('RedBalance', ''), 'From CR2')
-        hdu.header.set('BLUEBAL', exif.get('BlueBalance', ''), 'From CR2')
-        hdu.header.set('WBRGGB', exif.get(
-            'WB RGGBLevelAsShot', ''), 'From CR2')
-        hdu.header.set('DATE-OBS', obs_date)
-
-        hdu.header.set('IMAGEID', headers.get('image_id', ''))
-        hdu.header.set('SEQID', headers.get('sequence_id', ''))
-        hdu.header.set('FIELD', headers.get('field_name', ''))
-        hdu.header.set('RA-MNT', headers.get('ra_mnt', ''), 'Degrees')
-        hdu.header.set('HA-MNT', headers.get('ha_mnt', ''), 'Degrees')
-        hdu.header.set('DEC-MNT', headers.get('dec_mnt', ''), 'Degrees')
-
-        # Explicity convert the equinox for FITS header
-        try:
-            equinox = float(headers['equinox'].value.replace('J', ''))
-        except KeyError:
-            equinox = ''
-
-        hdu.header.set('EQUINOX', equinox)
-
-        hdu.header.set('AIRMASS', headers.get('airmass', ''), 'Sec(z)')
-        hdu.header.set('FILTER', headers.get('filter', ''))
-        hdu.header.set('LAT-OBS', headers.get('latitude', ''), 'Degrees')
-        hdu.header.set('LONG-OBS', headers.get('longitude', ''), 'Degrees')
-        hdu.header.set('ELEV-OBS', headers.get('elevation', ''), 'Meters')
-        hdu.header.set('MOONSEP', headers.get(
-            'moon_separation', ''), 'Degrees')
-        hdu.header.set('MOONFRAC', headers.get('moon_fraction', ''))
-        hdu.header.set('CREATOR', headers.get(
-            'creator', ''), 'POCS Software version')
-        hdu.header.set('INSTRUME', headers.get('camera_uid', ''), 'Camera ID')
-        hdu.header.set('OBSERVER', headers.get(
-            'observer', ''), 'PANOPTES Unit ID')
-        hdu.header.set('ORIGIN', headers.get('origin', ''))
-        hdu.header.set(
-            'RA-RATE', headers.get('tracking_rate_ra', ''), 'RA Tracking Rate')
-
-        if verbose:
-            print("Adding provided FITS header")
-
-        for key, value in fits_headers.items():
-            try:
-                hdu.header.set(key.upper()[0: 8], value)
-            except Exception:
-                pass
-
-        try:
-            if verbose:
-                print("Saving fits file to: {}".format(fits_fname))
-
-            hdu.writeto(fits_fname, output_verify='silentfix', clobber=clobber)
-        except Exception as e:
-            warn("Problem writing FITS file: {}".format(e))
-        else:
-            if remove_cr2:
-                os.unlink(cr2_fname)
-
-    return fits_fname
-
-
-def cr2_to_pgm(
-        cr2_fname,
-        pgm_fname=None,
-        dcraw='dcraw',
-        clobber=True, *args,
-        **kwargs):  # pragma: no cover
-    """ Convert CR2 file to PGM
-
-    Converts a raw Canon CR2 file to a netpbm PGM file via `dcraw`. Assumes
-    `dcraw` is installed on the system
-
-    Note:
-        This is a blocking call
-
-    Arguments:
-        cr2_fname {str} -- Name of CR2 file to convert
-        **kwargs {dict} -- Additional keywords to pass to script
-
-    Keyword Arguments:
-        pgm_fname {str} -- Name of PGM file to output, if None (default) then
-                           use same name as CR2 (default: {None})
-        dcraw {str} -- Path to installed `dcraw` (default: {'dcraw'})
-        clobber {bool} -- A bool indicating if existing PGM should be clobbered
-                         (default: {True})
-
-    Returns:
-        str -- Filename of PGM that was created
-
-    """
-
-    assert subprocess.call('dcraw', stdout=subprocess.PIPE),\
-        "could not execute dcraw in path: {}".format(dcraw)
-    assert os.path.exists(cr2_fname), "cr2 file does not exist at {}".format(
-                                      cr2_fname)
-
-    verbose = kwargs.get('verbose', False)
-
-    if pgm_fname is None:
-        pgm_fname = cr2_fname.replace('.cr2', '.pgm')
-
-    if os.path.exists(pgm_fname) and not clobber:
-        if verbose:
-            print("PGM file exists, returning existing file: {}".format(
-                  pgm_fname))
-    else:
-        try:
-            # Build the command for this file
-            command = '{} -t 0 -D -4 {}'.format(dcraw, cr2_fname)
-            cmd_list = command.split()
-            if verbose:
-                print("PGM Conversion command: \n {}".format(cmd_list))
-
-            # Run the command
-            if subprocess.check_call(cmd_list) == 0:
-                if verbose:
-                    print("PGM Conversion command successful")
-
-        except subprocess.CalledProcessError as err:
-            raise error.InvalidSystemCommand(msg="File: {} \n err: {}".format(
-                cr2_fname, err))
-
-    return pgm_fname
-
-
-def read_exif(fname, exiftool='exiftool'):  # pragma: no cover
-    """ Read the EXIF information
-
-    Gets the EXIF information using exiftool
-
-    Note:
-        Assumes the `exiftool` is installed
-
-    Args:
-        fname {str} -- Name of file (CR2) to read
-
-    Keyword Args:
-        exiftool {str} -- Location of exiftool (default: {'/usr/bin/exiftool'})
-
-    Returns:
-        dict -- Dictonary of EXIF information
-
-    """
-    assert os.path.exists(fname), warn("File does not exist: {}".format(fname))
-    exif = {}
-
-    try:
-        # Build the command for this file
-        command = '{} -j {}'.format(exiftool, fname)
-        cmd_list = command.split()
-
-        # Run the command
-        exif = loads(subprocess.check_output(cmd_list).decode('utf-8'))
-    except subprocess.CalledProcessError as err:
-        raise error.InvalidSystemCommand(
-            msg="File: {} \n err: {}".format(fname, err))
-
-    return exif[0]
-
-
-def read_pgm(fname, byteorder='>', remove_after=False):  # pragma: no cover
-    """Return image data from a raw PGM file as numpy array.
-
-    Note:
-        Format Spec: http://netpbm.sourceforge.net/doc/pgm.html
-        Source: http://stackoverflow.com/questions/7368739/numpy-and-16-bit-pgm
-
-    Note:
-        This is correctly processed as a Big endian even though the CR2 itself
-        marks it as a Little endian. See the notes in Source page above as well
-        as the comment about significant bit in the Format Spec
-
-    Args:
-        fname(str):         Filename of PGM to be converted
-        byteorder(str):     Big endian
-        remove_after(bool): Delete fname file after reading, defaults to False.
-        clobber(bool):      Clobber existing PGM or not, defaults to True
-
-    Returns:
-        numpy.array:        The raw data from the PGMx
-
-    """
-
-    with open(fname, 'rb') as f:
-        buffer = f.read()
-
-    # We know our header info is 19 chars long
-    header_offset = 19
-
-    img_type, img_size, img_max_value, _ = buffer[
-        0:header_offset].decode().split('\n')
-
-    assert img_type == 'P5', warn("No a PGM file")
-
-    # Get the width and height (as strings)
-    width, height = img_size.split(' ')
-
-    data = np.flipud(np.frombuffer(buffer[header_offset:],
-                                   dtype=byteorder + 'u2',
-                                   ).reshape((int(height), int(width))))
-
-    if remove_after:
-        os.remove(fname)
-
-    return data
-
-
-def create_timelapse(directory, fn_out=None, **kwargs):  # pragma: no cover
-    """Create a timelapse
-
-    A timelapse is created from all the jpg images in a given `directory`
-
-    Args:
-        directory (str): Directory containing jpg files
-        fn_out (str, optional): Full path to output file name, if not provided,
-            defaults to `directory` basename.
-        **kwargs (dict): Valid keywords: verbose
-
-    Returns:
-        str: Name of output file
-    """
-    if fn_out is None:
-        head, tail = os.path.split(directory)
-        if tail is '':
-            head, tail = os.path.split(head)
-
-        field_name = head.split('/')[-2]
-        cam_name = head.split('/')[-1]
-        fn_out = '{}/images/timelapse/{}_{}_{}.mp4'.format(
-            os.getenv('PANDIR'), field_name, cam_name, tail)
-
-    try:
-        ff = FFmpeg(
-            global_options='-r 3 -pattern_type glob',
-            inputs={'{}/*.jpg'.format(directory): None},
-            outputs={fn_out: '-s hd1080 -vcodec libx264'}
-        )
-
-        if 'verbose' in kwargs:
-            out = None
-            err = None
-            print("Timelapse command: ", ff.cmd)
-        else:
-            out = open(os.devnull, 'w')
-            err = open(os.devnull, 'w')
-
-        ff.run(stdout=out, stderr=err)
-    except Exception as e:
-        warn("Problem creating timelapse: {}".format(fn_out))
-        fn_out = None
-
-    return fn_out
-
-
-def clean_observation_dir(dir_name, *args, **kwargs):
-    """ Clean an observation directory
-
-    For the given `dir_name`, will:
-        * Compress FITS files
-        * Remove `.solved` files
-        * Create timelapse from JPG files if present
-        * Remove JPG files
-
-    Args:
-        dir_name (str): Full path to observation directory
-        *args: Description
-        **kwargs: Can include `verbose`
-    """
-    verbose = kwargs.get('verbose', False)
-
-    if verbose:
-        print("Cleaning dir: {}".format(dir_name))
-
-    # Pack the fits filts
-    try:
-        print("Packing FITS files")
-        for f in glob('{}/*.fits'.format(dir_name)):
-            try:
-                fpack(f)
-            except Exception as e:
-                warn(
-                    'Could not compress fits file: {}'.format(e))
-    except Exception as e:
-        warn(
-            'Problem with cleanup cleaning FITS:'.format(e))
-
-    try:
-        # Remove .solved files
-        print('Removing .solved files')
-        for f in glob('{}/*.solved'.format(dir_name)):
-            try:
-                os.remove(f)
-            except OSError as e:
-                warn(
-                    'Could not delete file: {}'.format(e))
-    except Exception as e:
-        warn(
-            'Problem with cleanup removing solved:'.format(e))
-
-    try:
-        jpg_list = glob('{}/*.jpg'.format(dir_name))
-
-        if len(jpg_list) > 0:
-
-            # Create timelapse
-            print(
-                'Creating timelapse for {}'.format(dir_name))
-            video_file = create_timelapse(dir_name)
-            print(
-                'Timelapse created: {}'.format(video_file))
-
-            # Remove jpgs
-            print('Removing jpgs')
-            for f in jpg_list:
-                try:
-                    os.remove(f)
-                except OSError as e:
-                    warn(
-                        'Could not delete file: {}'.format(e))
-    except Exception as e:
-        warn(
-            'Problem with cleanup creating timelapse:'.format(e))
-
-
-def analyze_polar_rotation(pole_fn, *args, **kwargs):
-    """ Get celestial pole XY coordinates
-
-    Args:
-        pole_fn (str): FITS file of celestial pole
-
-    Returns:
-        tuple(int): A tuple of integers corresponding to the XY pixel position
-        of celestial pole
-    """
-    get_solve_field(pole_fn, **kwargs)
-
-    wcs = WCS(pole_fn)
-
-    pole_cx, pole_cy = wcs.all_world2pix(360, 90, 1)
-
-    return pole_cx, pole_cy
-
-
-def analyze_ra_rotation(rotate_fn):
-    """ Get RA axis center of rotation XY coordinates
-
-    Args:
-        rotate_fn (str): FITS file of RA rotation
-
-    Returns:
-        tuple(int): A tuple of integers corresponding to the XY pixel position
-        of the center of rotation around RA axis
-    """
-    d0 = fits.getdata(rotate_fn)  # - 2048
-
-    # Get center
-    position = (d0.shape[1] // 2, d0.shape[0] // 2)
-    size = (1500, 1500)
-    d1 = Cutout2D(d0, position, size)
-
-    d1.data = d1.data / d1.data.max()
-
-    # Get edges for rotation
-    rotate_edges = canny(d1.data, sigma=1.0)
-
-    rotate_hough_radii = np.arange(100, 500, 50)
-    rotate_hough_res = hough_circle(rotate_edges, rotate_hough_radii)
-    rotate_accums, rotate_cx, rotate_cy, rotate_radii = \
-        hough_circle_peaks(rotate_hough_res, rotate_hough_radii, total_num_peaks=1)
-
-    return d1.to_original_position((rotate_cx[-1], rotate_cy[-1]))
-
-
-def plot_center(pole_fn, rotate_fn, pole_center, rotate_center):
-    """ Overlay the celestial pole and RA rotation axis images
-
-    Args:
-        pole_fn (str): FITS file of polar center
-        rotate_fn (str): FITS file of RA rotation image
-        pole_center (tuple(int)): Polar center XY coordinates
-        rotate_center (tuple(int)): RA axis center of rotation XY coordinates
-
-    Returns:
-        matplotlib.Figure: Plotted image
-    """
-    d0 = fits.getdata(pole_fn) - 2048.  # Easy cast to float
-    d1 = fits.getdata(rotate_fn) - 2048.  # Easy cast to float
-
-    d0 /= d0.max()
-    d1 /= d1.max()
-
-    pole_cx, pole_cy = pole_center
-    rotate_cx, rotate_cy = rotate_center
-
-    d_x = pole_center[0] - rotate_center[0]
-    d_y = pole_center[1] - rotate_center[1]
-
-    fig, ax = plt.subplots(ncols=1, nrows=1, figsize=(20, 14))
-
-    # Show rotation center in red
-    ax.scatter(rotate_cx, rotate_cy, color='r', marker='x', lw=5)
-
-    # Show polar center in green
-    ax.scatter(pole_cx, pole_cy, color='g', marker='x', lw=5)
-
-    # Show both images in background
-    norm = ImageNormalize(stretch=SqrtStretch())
-    ax.imshow(d0 + d1, cmap='Greys_r', norm=norm, origin='lower')
-
-    # Show an arrow
-    if (np.abs(pole_cy - rotate_cy) > 25) or (np.abs(pole_cx - rotate_cx) > 25):
-        ax.arrow(rotate_cx, rotate_cy, pole_cx - rotate_cx, pole_cy -
-                 rotate_cy, fc='r', ec='r', width=20, length_includes_head=True)
-
-    ax.set_title("dx: {:0.2f} pix      dy: {:0.2f} pix".format(d_x, d_y))
-
-    return fig
