+++ pocs/utils/images/fits.py
@@ -1,0 +1,322 @@
+import os
+import shutil
+import subprocess
+
+from warnings import warn
+
+from astropy.io import fits
+from astropy.wcs import WCS
+from astropy import units as u
+
+from pocs.utils import error
+
+
+def solve_field(fname, timeout=15, solve_opts=None, **kwargs):
+    """ Plate solves an image.
+
+    Args:
+        fname(str, required):       Filename to solve in .fits extension.
+        timeout(int, optional):     Timeout for the solve-field command,
+                                    defaults to 60 seconds.
+        solve_opts(list, optional): List of options for solve-field.
+        verbose(bool, optional):    Show output, defaults to False.
+    """
+    verbose = kwargs.get('verbose', False)
+    if verbose:
+        print("Entering solve_field")
+
+    solve_field_script = "{}/scripts/solve_field.sh".format(
+        os.getenv('POCS'), '/var/panoptes/POCS')
+
+    if not os.path.exists(solve_field_script):  # pragma: no cover
+        raise error.InvalidSystemCommand(
+            "Can't find solve-field: {}".format(solve_field_script))
+
+    # Add the options for solving the field
+    if solve_opts is not None:
+        options = solve_opts
+    else:
+        options = [
+            '--guess-scale',
+            '--cpulimit', str(timeout),
+            '--no-verify',
+            '--no-plots',
+            '--crpix-center',
+            '--match', 'none',
+            '--corr', 'none',
+            '--wcs', 'none',
+            '--downsample', '4',
+        ]
+
+        if kwargs.get('clobber', True):
+            options.append('--overwrite')
+        if kwargs.get('skip_solved', True):
+            options.append('--skip-solved')
+
+        if 'ra' in kwargs:
+            options.append('--ra')
+            options.append(str(kwargs.get('ra')))
+        if 'dec' in kwargs:
+            options.append('--dec')
+            options.append(str(kwargs.get('dec')))
+        if 'radius' in kwargs:
+            options.append('--radius')
+            options.append(str(kwargs.get('radius')))
+
+    cmd = [solve_field_script] + options + [fname]
+    if verbose:
+        print("Cmd:", cmd)
+
+    try:
+        proc = subprocess.Popen(cmd, universal_newlines=True,
+                                stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+    except OSError as e:
+        raise error.InvalidCommand(
+            "Can't send command to solve_field.sh: {} \t {}".format(e, cmd))
+    except ValueError as e:
+        raise error.InvalidCommand(
+            "Bad parameters to solve_field: {} \t {}".format(e, cmd))
+    except Exception as e:
+        raise error.PanError("Timeout on plate solving: {}".format(e))
+
+    if verbose:
+        print("Returning proc from solve_field")
+
+    return proc
+
+
+def get_solve_field(fname, replace=True, remove_extras=True, **kwargs):
+    """Convenience function to wait for `solve_field` to finish.
+
+    This function merely passes the `fname` of the image to be solved along to `solve_field`,
+    which returns a subprocess.Popen object. This function then waits for that command
+    to complete, populates a dictonary with the EXIF informaiton and returns. This is often
+    more useful than the raw `solve_field` function
+
+    Args:
+        fname ({str}): Name of FITS file to be solved
+        replace (bool, optional): Replace fname the solved file
+        remove_extras (bool, optional): Remove the files generated by solver
+        **kwargs ({dict}): Options to pass to `solve_field`
+
+    Returns:
+        dict: Keyword information from the solved field
+    """
+    verbose = kwargs.get('verbose', False)
+    out_dict = {}
+    output = None
+    errs = None
+
+    # Check for solved file
+    if kwargs.get('skip_solved', True) and \
+            (os.path.exists(fname.replace('.fits', '.solved')) or WCS(fname).is_celestial):
+        if verbose:
+            print("Solved file exists, skipping",
+                  "(pass skip_solved=False to solve again):",
+                  fname)
+
+        out_dict['solved_fits_file'] = fname
+        return out_dict
+
+    if verbose:
+        print("Entering get_solve_field:", fname)
+
+    # Set a default radius of 15
+    kwargs.setdefault('radius', 15)
+
+    proc = solve_field(fname, **kwargs)
+    try:
+        output, errs = proc.communicate(timeout=kwargs.get('timeout', 30))
+    except subprocess.TimeoutExpired:
+        proc.kill()
+        raise error.Timeout("Timeout while solving")
+    else:
+        if verbose:
+            print("Returncode:", proc.returncode)
+            print("Output:", output)
+            print("Errors:", errs)
+
+        if proc.returncode == 3:
+            raise error.SolveError('solve-field not found: {}'.format(output))
+
+        if not os.path.exists(fname.replace('.fits', '.solved')):
+            raise error.SolveError('File not solved')
+
+        try:
+            # Handle extra files created by astrometry.net
+            new = fname.replace('.fits', '.new')
+            rdls = fname.replace('.fits', '.rdls')
+            axy = fname.replace('.fits', '.axy')
+            xyls = fname.replace('.fits', '-indx.xyls')
+
+            if replace and os.path.exists(new):
+                # Remove converted fits
+                os.remove(fname)
+                # Rename solved fits to proper extension
+                os.rename(new, fname)
+
+                out_dict['solved_fits_file'] = fname
+            else:
+                out_dict['solved_fits_file'] = new
+
+            if remove_extras:
+                for f in [rdls, xyls, axy]:
+                    if os.path.exists(f):
+                        os.remove(f)
+
+        except Exception as e:
+            warn('Cannot remove extra files: {}'.format(e))
+
+    if errs is not None:
+        warn("Error in solving: {}".format(errs))
+    else:
+
+        try:
+            out_dict.update(fits.getheader(fname))
+        except OSError:
+            if verbose:
+                print("Can't read fits header for:", fname)
+
+    return out_dict
+
+
+def get_wcsinfo(fits_fname, verbose=False):
+    """Returns the WCS information for a FITS file.
+
+    Uses the `wcsinfo` astrometry.net utility script to get the WCS information
+    from a plate-solved file.
+
+    Parameters
+    ----------
+    fits_fname : {str}
+        Name of a FITS file that contains a WCS.
+    verbose : {bool}, optional
+        Verbose (the default is False)
+    Returns
+    -------
+    dict
+        Output as returned from `wcsinfo`
+    """
+    assert os.path.exists(fits_fname), warn(
+        "No file exists at: {}".format(fits_fname))
+
+    wcsinfo = shutil.which('wcsinfo')
+    if wcsinfo is None:
+        raise error.InvalidCommand('wcsinfo not found')
+
+    run_cmd = [wcsinfo, fits_fname]
+
+    if verbose:
+        print("wcsinfo command: {}".format(run_cmd))
+
+    proc = subprocess.Popen(run_cmd, stdout=subprocess.PIPE,
+                            stderr=subprocess.STDOUT, universal_newlines=True)
+    try:
+        output, errs = proc.communicate(timeout=5)
+    except subprocess.TimeoutExpired:  # pragma: no cover
+        proc.kill()
+        output, errs = proc.communicate()
+
+    unit_lookup = {
+        'crpix0': u.pixel,
+        'crpix1': u.pixel,
+        'crval0': u.degree,
+        'crval1': u.degree,
+        'cd11': (u.deg / u.pixel),
+        'cd12': (u.deg / u.pixel),
+        'cd21': (u.deg / u.pixel),
+        'cd22': (u.deg / u.pixel),
+        'imagew': u.pixel,
+        'imageh': u.pixel,
+        'pixscale': (u.arcsec / u.pixel),
+        'orientation': u.degree,
+        'ra_center': u.degree,
+        'dec_center': u.degree,
+        'orientation_center': u.degree,
+        'ra_center_h': u.hourangle,
+        'ra_center_m': u.minute,
+        'ra_center_s': u.second,
+        'dec_center_d': u.degree,
+        'dec_center_m': u.minute,
+        'dec_center_s': u.second,
+        'fieldarea': (u.degree * u.degree),
+        'fieldw': u.degree,
+        'fieldh': u.degree,
+        'decmin': u.degree,
+        'decmax': u.degree,
+        'ramin': u.degree,
+        'ramax': u.degree,
+        'ra_min_merc': u.degree,
+        'ra_max_merc': u.degree,
+        'dec_min_merc': u.degree,
+        'dec_max_merc': u.degree,
+        'merc_diff': u.degree,
+    }
+
+    wcs_info = {}
+    for line in output.split('\n'):
+        try:
+            k, v = line.split(' ')
+            try:
+                v = float(v)
+            except Exception:
+                pass
+
+            wcs_info[k] = float(v) * unit_lookup.get(k, 1)
+        except ValueError:
+            pass
+            # print("Error on line: {}".format(line))
+
+    wcs_info['wcs_file'] = fits_fname
+
+    return wcs_info
+
+
+def fpack(fits_fname, unpack=False, verbose=False):
+    """ Compress/Decompress a FITS file
+
+    Uses `fpack` (or `funpack` if `unpack=True`) to compress a FITS file
+
+    Parameters
+    ----------
+    fits_fname : {str}
+        Name of a FITS file that contains a WCS.
+    unpack : {bool}, optional
+        file should decompressed instead of compressed (default is False)
+    verbose : {bool}, optional
+        Verbose (the default is False)
+    Returns
+    -------
+    str
+        Filename of compressed/decompressed file
+    """
+    assert os.path.exists(fits_fname), warn(
+        "No file exists at: {}".format(fits_fname))
+
+    if unpack:
+        fpack = shutil.which('funpack')
+        run_cmd = [fpack, '-D', fits_fname]
+        out_file = fits_fname.replace('.fz', '')
+    else:
+        fpack = shutil.which('fpack')
+        run_cmd = [fpack, '-D', '-Y', fits_fname]
+        out_file = fits_fname.replace('.fits', '.fits.fz')
+
+    try:
+        assert fpack is not None
+    except AssertionError:
+        warn("fpack not found (try installing cfitsio). File has not been changed")
+        return fits_fname
+
+    if verbose:
+        print("fpack command: {}".format(run_cmd))
+
+    proc = subprocess.Popen(run_cmd, stdout=subprocess.PIPE,
+                            stderr=subprocess.STDOUT, universal_newlines=True)
+    try:
+        output, errs = proc.communicate(timeout=5)
+    except subprocess.TimeoutExpired:
+        proc.kill()
+        output, errs = proc.communicate()
+
+    return out_file
